% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magicBatch.R
\name{magicBatch}
\alias{magicBatch}
\title{Impute data using the MAGIC algorithm}
\usage{
magicBatch(
  data,
  select_features = NULL,
  mar_mat_input = NULL,
  import_mar_mat = FALSE,
  use_numpy = TRUE,
  pca = 20,
  t_param = c(2, 4, 6),
  n_diffusion_components = 0,
  k = 9,
  ka = 3,
  epsilon = 1,
  rescale_percent = 90,
  rescale_method = "adaptive",
  python_command = system("which python3", intern = TRUE)
)
}
\arguments{
\item{data}{An expression matrix where cells correspond to rows and genes correspond to columns}

\item{select_features}{A vector of features to use for imputation.}

\item{mar_mat_input}{A matrix where cells correspond to rows and components or features correspond to columns. If left unspecified, the Marcov matrix calculation is initialized with PCA of data.}

\item{import_mar_mat}{Whether to return the Markov matrix}

\item{pca}{An integer specifying the number of PCA components that should be used}

\item{t_param}{An integer or a vector of integers to be used to power the marcov affinity matrix}

\item{n_diffusion_components}{Number of diffusion map components to compute. If set to 0, this diffusion map will not be computed.}

\item{k}{The number of nearest neighbors used to construct the knn graph}

\item{ka}{This controls the standard deviation used in the Gaussian kernel width for a given cell, which is set to the distance to the ka-th nearest neighbor.}

\item{epsilon}{Epsilon parameter used in MAGIC}

\item{rescale_percent}{Percentile to rescale data to after imputation}

\item{rescale_method}{A string passed to the rescale_method argument of the rescale_data function. Two methods are available: "adaptive" or "classic" See rescale_data function for details.}

\item{python_command}{A character string passed to the "command" argument of the system2 function in order to invoke python. E.g. "/usr/local/bin/python3" on a Mac.}
}
\value{
A \code{list} that includes the following elements:
\describe{
  \item{imputed_data}{A cell by gene \code{matrix} of the imputed gene expression values.}
  \item{diffusion_map}{A cell by diffusion map component \code{matrix}.}
  \item{marcov_matrix}{A cell by cell \code{matrix} of the unpowered markov affinity matrix.}
}
}
\description{
This is an implementation of the Marcov Affinity-based Graph Imputation of Cells (MAGIC) algorithm described in 
Van Dijk, David et al, modified for speed and flexibility. If the mar_mat_input argument is used, the diffusion operator is computed independently of the data to be imputed. 
This allows any low-dimensional representation of the data, including batch-corrected data, to be directly used to calculate the powered Marcov 
affinity matrix.
}
\author{
Kevin Brulois
}
