% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magicBatch.R
\name{multi_t_fast_impute}
\alias{multi_t_fast_impute}
\title{Impute data using the MAGIC algorithm}
\usage{
multi_t_fast_impute(data, L, t_param)
}
\arguments{
\item{data}{An expression matrix where cells correspond to rows and genes correspond to columns}

\item{t_param}{A vector of integers ranging from 1 to 32}

\item{marcov_mat}{A Markov matrix in dgCMatrix format where cells correspond to rows and to columns}
}
\value{
A \code{list} that includes the following elements:
\describe{
  \item{imputed_data}{A cell by gene \code{matrix} of the imputed gene expression values.}
  \item{diffusion_map}{A cell by diffusion map component \code{matrix}.}
  \item{affinity_matrix}{A cell by cell \code{matrix} of the unpowered markov affinity matrix.}
}
}
\description{
This is a helper function that imputes data using a pre-computed Marcov matrix. It minimizing the number of matrix multiplications needed to 
compute a given set of matrix powering operations. The exponents (t_param) must be integers from 1 to 36.
}
\author{
Kevin Brulois
}
