% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magicBatch.R
\name{rescale_data}
\alias{rescale_data}
\title{Rescale Data}
\usage{
rescale_data(data, imputed_data, rescale_percent, rescale_method)
}
\arguments{
\item{data}{The original unimputed expression matrix where cells correspond to rows and genes correspond to columns}

\item{imputed_data}{Unscaled imputed data where cells corrsponnd to rows and genes correspond to columns}

\item{rescale_percent}{Percentile used to rescale imputed data.}

\item{rescale_method}{A string specifying one of three rescale methods: "adaptive" or "classic" The "classic" method performs rescaling as in the original version of MAGIC: rescaling factors are computed using the given percentile of the original and imputed data or the max if the data is zero at the given percentile. The "adaptive" method adjusts the rescale percentile for each gene between recale_percent and 100 based on the number of non-zero data points.}
}
\value{
A cell by gene \code{matrix} with the rescaled imputed gene expression values.
}
\description{
This function rescales imputed data using one of two methods. The "classic" method rescales the imputed data as in the original implementation.
The "adaptive" method adjusts rescaling factors for each gene using only cells with a non-zero expression value.
}
\author{
Kevin Brulois
}
